perf(hyperspace): fp16 towers, sort-based median, query-scale caching
- SigLIP2 dtype defaults to float16: 4.6 -> 2.3 GB VRAM, 85 -> 62 ms per
  frame, patch maps visually identical to fp32 (max score delta 0.06 on a
  0.34 range). The hyperspace text tower pins float32 when it runs on CPU.
- Sort-based per-dim median and a cached normalized aggregate: nanmedian
  plus per-prompt renormalization dominated query/save time on 300k+
  voxel maps and blew the 120 s save_map RPC deadline.
- tool_hyperspace saves the map after the queries so the first query
  pre-warms the aggregation cache, and runs replay at --speed 1 default
  workloads can sustain (2x replay starved the pipeline to 36/531 frames).

+++ dimos/models/embedding/siglip2.py
@@ -35,7 +35,9 @@ from dimos.msgs.sensor_msgs.Image import Image
 
 class SigLIP2ModelConfig(HuggingFaceEmbeddingModelConfig):
     model_name: str = "google/siglip2-base-patch16-256"
-    dtype: torch.dtype = torch.float32
+    #: fp16 halves VRAM (so400m: 4.6 -> 2.3 GB) at no visible quality cost.
+    #: Pass float32 explicitly for CPU inference — fp16 has no fast CPU path.
+    dtype: torch.dtype = torch.float16
     #: SigLIP text towers are trained on fixed-length padded input (64 tokens).
     text_max_length: int = 64
     #: Run each patch token through the vision attention-pooling head so patch
